Delegate user resolution for authorization request to an event listener
